Converted ingredient list to TreeSet. Got all Drinks loaded in

+++ src/main/java/ElBell/CocktailSpringBoot/utilities/DrinkReference.java
@@ -11,19 +11,19 @@ import java.net.URL;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class DrinkReference {
 
-    private Long id;
+    private Integer id;
     private URL thumb;
     private String name;
 
     public DrinkReference(@JsonProperty("strDrink") String name,
                           @JsonProperty("strDrinkThumb") URL thumb,
-                          @JsonProperty("idDrink") Long id) {
+                          @JsonProperty("idDrink") Integer id) {
         this.name = name;
         this.thumb = thumb;
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 }
